% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_emailjsr.R
\name{use_emailjsr_server}
\alias{use_emailjsr_server}
\title{use_emailjsr_server}
\usage{
use_emailjsr_server(
  id,
  service_id,
  user_id,
  template_id,
  access_token,
  language = "en"
)
}
\arguments{
\item{id}{Should be same to id of use_emailjsr_ui}

\item{service_id}{emailjs.com Service Id}

\item{user_id}{emailjs.com User Id}

\item{template_id}{emailjs.com Template Id}

\item{access_token}{emailjs.com Access Token}

\item{language}{Language of modal. "en" for English, and "ko" for Korean.}
}
\value{
{No return value, for shiny support}
}
\description{
emailjs.com R support with preset modal
}
\examples{
library(shiny)
library(emailjsr)
ui <- fluidPage(
  use_emailjsr_ui("id")
)

server <- function(input, output, session) {
  emailjsr::use_emailjsr_server("id",
    service_id = "Your_service_id",
    template_id = "Your_template_id",
    user_id = "Your_user_id",
    access_token = "Your_Access_Token",
    language = "en"
  )
}
}
