% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailjsr.R
\name{send_email}
\alias{send_email}
\title{send_email}
\usage{
send_email(
  service_id,
  user_id,
  template_id,
  template_params = list(),
  access_token = NULL
)
}
\arguments{
\item{service_id}{emailjs.com Service Id}

\item{user_id}{emailjs.com User Id}

\item{template_id}{emailjs.com Template Id}

\item{template_params}{Params passed to emailjs.com}

\item{access_token}{emailjs.com Access Token. Default is NULL in case of you don't use access token.}
}
\description{
emailjs.com R support
}
\examples{
library(emailjsr)

emailjsr::send_email("service_id", "user_id", "template_id", "access_token", list(name = "John Doe"))

}
