% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_emailjsr.R
\name{use_emailjsr}
\alias{use_emailjsr}
\title{use_emailjsr}
\usage{
use_emailjsr(serviceId = NULL, userId = NULL, templateId = NULL)
}
\arguments{
\item{serviceId}{emailjs.com Service Id}

\item{userId}{emailjs.com User Id}

\item{templateId}{emailjs.com Template Id}
}
\description{
emailjs.com R support
}
\examples{
library(emailjsr)

}
