% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_emailjsr.R
\name{use_emailjsr_ui}
\alias{use_emailjsr_ui}
\title{use_emailjsr}
\usage{
use_emailjsr_ui(id)
}
\arguments{
\item{id}{Should be same to id of use_emailjsr_server}

\item{serviceId}{emailjs.com Service Id}

\item{userId}{emailjs.com User Id}

\item{templateId}{emailjs.com Template Id}

\item{access_token}{emailjs.com Access Token}
}
\description{
emailjs.com R support
}
\examples{
library(emailjsr)

}
