% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_emailjsr.R, R/use_emailjsr2.R
\name{use_emailjsr_ui}
\alias{use_emailjsr_ui}
\alias{use_emailjsr_ui2}
\title{use_emailjsr_ui}
\usage{
use_emailjsr_ui(id, message = "Show feedback modal")

use_emailjsr_ui2(id, message = "Show feedback modal")
}
\arguments{
\item{id}{Should be same to id of use_emailjsr_server.}

\item{message}{Messege on button.}
}
\description{
emailjs.com R support with preset modal

emailjs.com R support
}
\examples{
library(shiny)
library(emailjsr)
ui <- fluidPage(
  use_emailjsr_ui("id")
)

server <- function(input, output, session) {
  emailjsr::use_emailjsr_server("id",
    service_id = "Your_service_id",
    template_id = "Your_template_id", user_id = "Your_user_id",
    access_token = "Your_Access_Token"
  )
}

library(shiny)
library(emailjsr)
ui <- fluidPage(
  use_emailjsr_ui("id")
)

server <- function(input, output, session) {
  emailjsr::use_emailjsr_server("id",
    service_id = "Your_service_id",
    template_id = "Your_template_id", user_id = "Your_user_id",
    access_token = "Your_Access_Token"
  )
}

}
